+++ Src/01/06/KSociety.Base.Infra.Shared/Class/Identity/RepositoryBase.cs
@@ -1,5 +1,5 @@
-﻿using KSociety.Base.Infra.Shared.Csv;
-using KSociety.Base.Infra.Shared.Interface.Identity;
+﻿using KSociety.Base.Infra.Shared.Interface.Identity;
+using KSociety.Base.InfraSub.Shared.Class.Csv;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.ChangeTracking;
