+++ Src/01/KSociety.Base.InfraSub.Shared/Class/Csv/ReadCsv.cs
@@ -1,5 +1,4 @@
 ﻿using CsvHelper;
-using CsvHelper.Configuration;
 using Microsoft.Extensions.Logging;
 using System;
 using System.Collections.Generic;
@@ -10,14 +9,13 @@ using System.Runtime.CompilerServices;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace KSociety.Base.Infra.Shared.Csv;
+namespace KSociety.Base.InfraSub.Shared.Class.Csv;
 
-public static class ReadCsvClassMap<TEntity, TClassMap>
-    where TEntity : class
-    where TClassMap : ClassMap<TEntity>
+public static class ReadCsv<TClass>
+    where TClass : class
 {
 
-    public static TEntity[] Read(ILoggerFactory loggerFactory, string fileName)
+    public static TClass[] Read(ILoggerFactory loggerFactory, string fileName)
     {
         var logger = loggerFactory?.CreateLogger("ReadCsv");
         var csvFileName = @"." + fileName + @".csv";
@@ -32,8 +30,8 @@ public static class ReadCsvClassMap<TEntity, TClassMap>
             using var stream = assembly.GetManifestResourceStream(resourceName);
             using var streamReader = new StreamReader(stream ?? throw new InvalidOperationException());
             var reader = new CsvReader(streamReader, Configuration.CsvConfiguration);
-            reader.Context.RegisterClassMap<TClassMap>();
-            return reader.GetRecords<TEntity>().ToArray();
+
+            return reader.GetRecords<TClass>().ToArray();
         }
         catch (Exception ex)
         {
@@ -42,17 +40,16 @@ public static class ReadCsvClassMap<TEntity, TClassMap>
         return null;
     }
 
-    public static IEnumerable<TEntity> Import(ILoggerFactory loggerFactory, string fileName)
+    public static IEnumerable<TClass> Import(ILoggerFactory loggerFactory, string fileName)
     {
         var logger = loggerFactory?.CreateLogger("ImportCsv");
-
+            
         try
         {
             using var streamReader = new StreamReader(fileName);
             var reader = new CsvReader(streamReader, Configuration.CsvConfiguration);
-            reader.Context.RegisterClassMap<TClassMap>();
 
-            return reader.GetRecords<TEntity>().ToArray();
+            return reader.GetRecords<TClass>().ToArray();
         }
         catch (Exception ex)
         {
@@ -61,7 +58,7 @@ public static class ReadCsvClassMap<TEntity, TClassMap>
         return null;
     }
 
-    public static IEnumerable<TEntity> Import(ILoggerFactory loggerFactory, byte[] byteArray)
+    public static IEnumerable<TClass> Import(ILoggerFactory loggerFactory, byte[] byteArray)
     {
         var logger = loggerFactory?.CreateLogger("ImportCsv");
 
@@ -69,8 +66,8 @@ public static class ReadCsvClassMap<TEntity, TClassMap>
         {
             using var streamReader = new StreamReader(new MemoryStream(byteArray));
             var reader = new CsvReader(streamReader, Configuration.CsvConfiguration);
-            reader.Context.RegisterClassMap<TClassMap>();
-            return reader.GetRecords<TEntity>().ToArray();
+
+            return reader.GetRecords<TClass>().ToArray();
         }
         catch (Exception ex)
         {
@@ -79,37 +76,34 @@ public static class ReadCsvClassMap<TEntity, TClassMap>
         return null;
     }
 
-    public static IAsyncEnumerable<TEntity> ImportAsync(ILoggerFactory loggerFactory, string fileName)
-    {
-        var logger = loggerFactory?.CreateLogger("ImportAsyncCsv");
-        IAsyncEnumerable<TEntity> output = null;
-
-        try
-        {
-            using var streamReader = new StreamReader(fileName);
-            var reader = new CsvReader(streamReader, Configuration.CsvConfiguration);
-            reader.Context.RegisterClassMap<TClassMap>();
-            output = reader.GetRecordsAsync<TEntity>();
-        }
-        catch (Exception ex)
-        {
-            logger?.LogError(ex, "ReadCsv.ImportAsync: ");
-        }
-        return output;
-    }
-
-    public static async IAsyncEnumerable<TEntity> ImportAsync(ILoggerFactory loggerFactory, byte[] byteArray, [EnumeratorCancellation] CancellationToken cancellationToken = default)
+    public static async IAsyncEnumerable<TClass> ImportAsync(ILoggerFactory loggerFactory, string fileName, [EnumeratorCancellation] CancellationToken cancellationToken = default)
     {
         var logger = loggerFactory?.CreateLogger("ImportAsyncCsv");
 
-        using var streamReader = new StreamReader(new MemoryStream(byteArray));
+        using var streamReader = new StreamReader(fileName);
         var reader = new CsvReader(streamReader, Configuration.CsvConfiguration);
-        reader.Context.RegisterClassMap<TClassMap>();
-        var result = reader.GetRecordsAsync<TEntity>(cancellationToken);
+
+        var result = reader.GetRecordsAsync<TClass>(cancellationToken);
 
         await foreach (var item in result.WithCancellation(cancellationToken).ConfigureAwait(false))
         {
             yield return item;
         }
+    }
+
+    public static async IAsyncEnumerable<TClass> ImportAsync(ILoggerFactory loggerFactory, byte[] byteArray, [EnumeratorCancellation] CancellationToken cancellationToken = default)
+    {
+        var logger = loggerFactory?.CreateLogger("ImportAsyncCsv");
+
+        using var streamReader = new StreamReader(new MemoryStream(byteArray));
+        var reader = new CsvReader(streamReader, Configuration.CsvConfiguration);
+
+        var result = reader.GetRecordsAsync<TClass>(cancellationToken);
+
+        await foreach (var item in result.WithCancellation(cancellationToken).ConfigureAwait(false))
+        {
+            yield return item;
+        }
+
     }
 }
