+++ Src/01/KSociety.Base.InfraSub.Shared/KSociety.Base.InfraSub.Shared.csproj
@@ -8,6 +8,7 @@
 	</PropertyGroup>
 
   <ItemGroup>   
+    <PackageReference Include="CsvHelper" Version="27.2.1" />   
     <PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.1.1">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
