+++ Src/01/Test/KSociety.Base.InfraSub.Test/CodeDomService.cs
@@ -23,7 +23,7 @@ namespace KSociety.Base.InfraSub.Test
 
             _codeDomService = new KSociety.Base.InfraSub.Shared.Class.CodeDom.CodeDomService();
 
-            _classGenerators = ReadCsvClassMap<KSociety.Base.InfraSub.Shared.Class.CodeDom.ClassGenerator, ClassMap.ClassGenerator>.Read(_loggerFactory, "Test");
+            _classGenerators = ReadCsvClassMap<KSociety.Base.InfraSub.Shared.Class.CodeDom.ClassGenerator, KSociety.Base.InfraSub.Shared.Class.CodeDom.ClassMap.ClassGenerator>.Read(_loggerFactory, "TestDto");
             ;
         }
