+++ Src/01/Test/KSociety.Base.InfraSub.Test/Csv/TestDto.csv
